feat: 🎸 modify network deposit
